requirements.txt
strawberry-graphql[fastapi]==0.240.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
